+++ Mailer/extensions/mailer-module/src/migrations/20230601081041_change_mail_template_link_url_to_text.php
@@ -1,0 +1,21 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class ChangeMailTemplateLinkUrlToText extends AbstractMigration
+{
+    public function up()
+    {
+        $this->table('mail_template_links')
+            ->changeColumn('url', 'text')
+            ->update();
+    }
+
+    public function down()
+    {
+        $this->table('mail_template_links')
+            ->changeColumn('url', 'string')
+            ->update();
+    }
+}